backend/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
openai
python-dotenv
yfinance
pillow
requests_html
azure-cosmos
python-dotenv
azure-ai-documentintelligence
azure-identity
PyMuPDF
langchain
langchain_core
langchain_community
langchain_openai
tiktoken
python-multipart
